local_deploy/api/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["AbiHanna-Guy <guyabihanna@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
celery = "^5.2.7"
fastapi = "^0.95.2"
pandas = "^2.0.2"
jupyter = "^1.0.0"
notebook = "^6.5.4"
hypercorn = "^0.14.3"
tortoise-orm = "^0.19.3"
sentry-sdk = "^1.24.0"
fastapi-socketio = "^0.0.10"
python-dotenv = "^1.0.0"
pydantic = "^1.10.8"
aiohttp = "^3.8.4"
python-socketio = "^5.8.0"
asyncpg = "^0.27.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
